requirements.txt
azure-cognitiveservices-vision-computervision==0.7.0
azure-common==1.1.28
azure-core==1.25.1
Pillow==8.4.0
python-dotenv==0.21.0
virtualenv==20.16.3
gunicorn==20.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dominate==2.5.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
requests>=2.24.0
SQLAlchemy==1.3.19
urllib3==1.25.10
